+++ src/test/java/dev/katsute/onemta/TestJson.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
 
-import static dev.katsute.onemta.Json.*;
+import static dev.katsute.onemta.Json2.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 @SuppressWarnings("SpellCheckingInspection")
